pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "my-toolkit"
version = "2.0.0"
dependencies = [
    "Pillow>=10.0.0",
    "Flask>=2.3.0",
    "Werkzeug>=2.3.0"
]

[project.scripts]
my-toolkit = "my_toolkit.__main__:main"